+++ src/main/java/com/burgosmanuel/portfolio/experiencia/ExperienciaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienciaRepository extends JpaRepository<Experiencia, Integer> {
+public interface ExperienciaRepository extends JpaRepository<Experiencia, Long> {
     
 }
