+++ src/main/java/com/burgosmanuel/portfolio/seccion/SeccionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeccionRepository extends JpaRepository<Seccion, Integer>{
+public interface SeccionRepository extends JpaRepository<Seccion, Long>{
     
 }
